+++ app/assets/javascripts/views/story_view.js
@@ -5,6 +5,8 @@ import StoryDescription from 'components/story/StoryDescription';
 import StoryHistoryLocation from 'components/story/StoryHistoryLocation';
 import StorySelect from 'components/story/StorySelect';
 import StoryDatePicker from 'components/story/StoryDatePicker';
+import StoryNotes from 'components/story/StoryNotes';
+import NoteForm from 'components/notes/NoteForm';
 
 var Clipboard = require('clipboard');
 
@@ -12,8 +14,6 @@ var executeAttachinary = require('libs/execute_attachinary');
 
 var FormView = require('./form_view');
 var EpicView = require('./epic_view');
-var NoteForm = require('./note_form');
-var NoteView = require('./note_view');
 var TaskForm = require('./task_form');
 var TaskView = require('./task_view');
 
@@ -32,8 +32,8 @@ module.exports = FormView.extend({
 
     _.bindAll(this, "render", "highlight", "moveColumn", "setClassName",
       "transition", "estimate", "disableForm", "renderNotes",
-      "renderNotesCollection", "addEmptyNote", "hoverBox",
-      "renderTasks", "renderTasksCollection", "addEmptyTask",
+      "hoverBox", "renderTasks", "handleNoteDelete", "handleSaveError",
+      "renderTasksCollection", "addEmptyTask", "handleNoteSubmit",
       "clickSave", "attachmentDone", "attachmentStart",
       "attachmentFail", "toggleControlButtons");
 
@@ -53,9 +53,6 @@ module.exports = FormView.extend({
     this.model.on("change:estimate", this.setClassName);
     this.model.on("change:state", this.setClassName);
 
-    this.model.on("change:notes", this.addEmptyNote);
-    this.model.on("change:notes", this.renderNotesCollection);
-
     this.model.on("change:tasks", this.addEmptyTask);
     this.model.on("change:tasks", this.renderTasksCollection);
 
@@ -73,8 +70,6 @@ module.exports = FormView.extend({
     // Set up CSS classes for the view
     this.setClassName();
 
-    // Add an empty note to the collection
-    this.addEmptyNote();
     // Add an empty task to the collection
     this.addEmptyTask();
   },
@@ -442,7 +437,8 @@ module.exports = FormView.extend({
 
       this.initTags();
 
-      this.renderNotes();
+      this.$el.append($('<div data-story-notes></div>'));
+      this.$el.append($('<div data-story-note-form></div>'));
 
       if(this.model.get('story_type') === 'release') {
         this.$el.empty();
@@ -568,6 +564,68 @@ module.exports = FormView.extend({
 
       this.bindElementToAttribute($storyRequestedBySelect.find('select[name="requested_by"]'), 'requested_by_id');
     }
+
+    this.renderNotes();
+  },
+
+  renderNotes: function() {
+    const $storyNotes = this.$('[data-story-notes]');
+    if ($storyNotes.length && !this.model.isNew()) {
+      const isReadonly = this.isReadonly();
+      const notes = this.model.notes;
+
+      ReactDOM.render(
+        <StoryNotes
+          notes={isReadonly ? notes : notes.slice(0,-1)}
+          disabled={isReadonly}
+          handleDelete={this.handleNoteDelete}
+        />,
+        $storyNotes.get(0)
+      );
+
+      if (!isReadonly) {
+        this.renderNoteForm();
+      }
+    }
+  },
+
+  renderNoteForm: function() {
+    const $noteForm = this.$('[data-story-note-form]');
+    if ($noteForm.length) {
+      this.addEmptyNote();
+
+      ReactDOM.render(
+        <NoteForm
+          note={this.model.notes.last()}
+          onSubmit={this.handleNoteSubmit}
+        />,
+        $noteForm.get(0)
+      );
+
+      $noteForm.find('textarea').atwho({
+        at: '@',
+        data: window.projectView.usernames()
+      });
+    }
+  },
+
+  handleNoteDelete: function(note) {
+    note.destroy();
+    this.renderNotes();
+  },
+
+  handleNoteSubmit: function({ note, newValue }) {
+    note.set({note: newValue});
+    return note.save(null, {
+      success: () => window.projectView.model.fetch(),
+      error: this.handleSaveError
+    });
+  },
+
+  handleSaveError: function(model, response) {
+    const json = JSON.parse(response.responseText);
+    model.set({errors: json[model.name].errors});
+    window.projectView.noticeSaveError(model);
   },
 
   createStoryEstimateOptions: function(option) {
@@ -715,15 +773,6 @@ module.exports = FormView.extend({
     });
   },
 
-  renderNotes: function() {
-    if (this.model.notes.length > 0) {
-      var el = this.$el;
-      el.append(this.label('notes', I18n.t('story.notes')));
-      el.append('<div class="notelist"/>');
-      this.renderNotesCollection();
-    }
-  },
-
   renderTasks: function() {
     if (this.model.tasks.length > 0) {
       var el = this.$el;
@@ -731,24 +780,6 @@ module.exports = FormView.extend({
       el.append('<div class="tasklist"/>');
       this.renderTasksCollection();
     }
-  },
-
-  renderNotesCollection: function() {
-    var notelist = this.$('div.notelist');
-    notelist.html('');
-    if(!this.isReadonly())
-      this.addEmptyNote();
-    var that = this;
-    this.model.notes.each(function(note) {
-      var view;
-      if (!that.isReadonly() && note.isNew()) {
-        view = new NoteForm({model: note});
-      } else {
-        if (that.isReadonly()) note.isReadonly = true;
-        view = new NoteView({model: note});
-      }
-      notelist.append(view.render().el);
-    });
   },
 
   renderTasksCollection: function() {