+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Story requested_by select to react component
 - Move select components to renderSelects function
 - Alter route button Manage Members Team and associate user in team
+- Story notes into react components
 
 ### Fixed
 - Story History bug when story actions where triggered